fetch refactoring
