+++ app/assets/javascripts/application.js
@@ -17,7 +17,3 @@
 //= require jquery_ujs
 //= require twitter/bootstrap
 //= require js-routes
-
-$(function() {
-  /* Your JavaScript goes here... */
-});
