+++ app/javascript/packs/components/EditPopup.js
@@ -2,7 +2,7 @@ import React from "react";
 import Select from "react-select";
 import UserSelect from "./UserSelect";
 import { Modal, Button, FormGroup, ControlLabel, FormControl } from "react-bootstrap";
-import { fetch } from "./Fetch";
+import FetchService from "../services/FetchService";
 
 export default class EditPopup extends React.Component {
   state = {
@@ -41,7 +41,7 @@ export default class EditPopup extends React.Component {
 
   loadCard = cardId => {
     this.setState({ isLoading: true });
-    fetch("GET", window.Routes.api_v1_task_path(cardId, { format: "json" })).then(({ data }) => {
+    FetchService.get(window.Routes.api_v1_task_path(cardId, { format: "json" })).then(({ data }) => {
       this.setState({ task: data });
       this.setState({ isLoading: false });
     });
@@ -62,7 +62,7 @@ export default class EditPopup extends React.Component {
   };
 
   handleCardEdit = () => {
-    fetch("PUT", window.Routes.api_v1_task_path(this.props.cardId, { format: "json" }), {
+    FetchService.put(window.Routes.api_v1_task_path(this.props.cardId, { format: "json" }), {
       name: this.state.task.name,
       description: this.state.task.description,
       author_id: this.state.task.author.id,
@@ -78,7 +78,7 @@ export default class EditPopup extends React.Component {
   };
 
   handleCardDelete = () => {
-    fetch("DELETE", window.Routes.api_v1_task_path(this.props.cardId, { format: "json" })).then(response => {
+    FetchService.delete(window.Routes.api_v1_task_path(this.props.cardId, { format: "json" })).then(response => {
       if (response.statusText == "OK") {
         this.props.onClose(this.state.task.state);
       } else {
