+++ app/javascript/packs/components/TaskBoard.js
@@ -4,7 +4,7 @@ import LaneHeader from "./LaneHeader";
 import AddPopup from "./AddPopup";
 import EditPopup from "./EditPopup";
 import axios from "axios";
-import { fetch } from "./Fetch";
+import FetchService from "../services/FetchService";
 import { Button } from "react-bootstrap";
 
 export default class TasksBoard extends React.Component {
@@ -79,8 +79,7 @@ export default class TasksBoard extends React.Component {
   }
 
   fetchLine(state, page = 1) {
-    return fetch(
-      "GET",
+    return FetchService.get(
       window.Routes.api_v1_tasks_path({ q: { state_eq: state }, page: page, per_page: 10, format: "json" })
     ).then(({ data }) => {
       return data;
@@ -100,7 +99,7 @@ export default class TasksBoard extends React.Component {
   };
 
   handleDragEnd = (cardId, sourceLaneId, targetLaneId) => {
-    fetch("PUT", window.Routes.api_v1_task_path(cardId, { format: "json" }), { task: { state: targetLaneId } }).then(() => {
+    FetchService.put(window.Routes.api_v1_task_path(cardId, { format: "json" }), { task: { state: targetLaneId } }).then(() => {
       this.loadLine(sourceLaneId);
       this.loadLine(targetLaneId);
     });
