+++ app/javascript/packs/components/UserSelect.js
@@ -1,32 +1,34 @@
 import React, { Component } from "react";
 import AsyncSelect from "react-select/lib/Async";
-import { fetch } from "./Fetch";
+import FetchService from "../services/FetchService";
+
 export default class UserSelect extends Component {
   state = {
     inputValue: ""
   };
-  getOptionLabel = option => {
-    return option.first_name + " " + option.last_name;
-  };
-  getOptionValue = option => {
-    return option.id;
-  };
-  loadOptions = inputValue => {
-    return fetch(
-      "GET",
+
+  getOptionLabel = ({ first_name, last_name }) => `${first_name} ${last_name}`;
+
+  getOptionValue = option => option.id;
+
+  loadOptions = () => {
+    const { inputValue } = this.state;
+    return FetchService.get(
       window.Routes.api_v1_users_path({ q: { first_name_or_last_name_cont: inputValue }, format: "json" })
     ).then(({ data }) => {
       return data.items;
     });
   };
+
   handleInputChange = newValue => {
     const inputValue = newValue.replace(/\W/g, "");
     this.setState({ inputValue });
-    return inputValue;
   };
+
   componentDidMount() {
     this.loadOptions();
   }
+
   render() {
     return (
       <div>
@@ -38,7 +40,6 @@ export default class UserSelect extends Component {
           getOptionLabel={this.getOptionLabel}
           getOptionValue={this.getOptionValue}
           isDisabled={this.props.isDisabled}
-          defaultValue={this.props.value}
           onChange={this.props.onChange}
         />
       </div>
