+++ app/views/layouts/application.html.slim
@@ -44,4 +44,4 @@ html lang="en"
 
       footer
         p
-        p &copy; Company 2018
+        p &copy; Company 2019
