+++ package.json
@@ -4,6 +4,7 @@
   "dependencies": {
     "@rails/webpacker": "^4.0.2",
     "axios": "^0.18.0",
+    "eslint-config-airbnb": "^17.1.0",
     "react": "^16.8.4",
     "react-bootstrap": "^0.32.3",
     "react-dom": "^16.8.4",
